requirements.txt
# Core dependencies
python-telegram-bot==20.7
openai==1.3.0
httpx==0.25.0
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1

# Additional dependencies
python-multipart==0.0.6
jinja2==3.1.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0


